+++ src/filters/edit.cpp
@@ -401,7 +401,7 @@ AVSValue __cdecl Splice::CreateAligned(AVSValue args, void*, IScriptEnvironment*
 AVSValue __cdecl Dissolve::Create(AVSValue args, void*, IScriptEnvironment* env) 
 {
   const int overlap = args[2].AsInt();
-  const float fps = args[3].AsFloat(24);
+  const float fps = (float)args[3].AsFloat(24);
   PClip result = args[0].AsClip();
   for (int i=0; i < args[1].ArraySize(); ++i)
     result = new Dissolve(result, args[1][i].AsClip(), overlap, fps, env);
@@ -541,7 +541,7 @@ void Dissolve::GetAudio(void* buf, __int64 start, __int64 count, IScriptEnvironm
     return;
   }
 
-  const int bytes = vi.BytesFromAudioSamples(count);
+  const int bytes = (int)vi.BytesFromAudioSamples(count);
   if (audbufsize < bytes) {
     delete[] audbuffer;
     audbuffer = new BYTE[bytes];
@@ -552,9 +552,9 @@ void Dissolve::GetAudio(void* buf, __int64 start, __int64 count, IScriptEnvironm
   child2->GetAudio(audbuffer, start - audio_fade_start, count, env);
   
   const int nch = vi.AudioChannels();
-  const int countXnch = count*nch;
-  const int denominator = (audio_fade_end - audio_fade_start);
-  int numerator = (audio_fade_end - start);
+  const int countXnch = (int)count*nch;
+  const int denominator = (int)(audio_fade_end - audio_fade_start);
+  int numerator = (int)(audio_fade_end - start);
   
   if (vi.IsSampleType(SAMPLE_INT16)) {
     short *const a = (short*)buf;
@@ -710,7 +710,7 @@ void Reverse::GetAudio(void* buf, __int64 start, __int64 count, IScriptEnvironme
   child->GetAudio(buf, vi.num_audio_samples - start - count, count, env);
   int xor = vi.BytesPerAudioSample() - 1;
   char* buf2 = (char*)buf;
-  const int count_bytes = vi.BytesFromAudioSamples(count);
+  const int count_bytes = (int)vi.BytesFromAudioSamples(count);
   for (int i=0; i<(count_bytes>>1); ++i) {
     char temp = buf2[i]; buf2[i] = buf2[count_bytes-1-(i^xor)]; buf2[count_bytes-1-(i^xor)] = temp;
   }
@@ -868,7 +868,7 @@ PClip __cdecl ColorClip(PClip a, int duration, int color, float fps, IScriptEnvi
 AVSValue __cdecl Create_FadeOut0(AVSValue args, void*,IScriptEnvironment* env) {
   const int duration = args[1].AsInt();
   const int fadeclr = args[2].AsInt(0);
-  const float fps = args[3].AsFloat(24);
+  const float fps = (float)args[3].AsFloat(24);
   PClip a = args[0].AsClip();
   PClip b = ColorClip(a,duration,fadeclr,fps,env);
   return new Dissolve(a, b, duration, fps, env);
@@ -877,7 +877,7 @@ AVSValue __cdecl Create_FadeOut0(AVSValue args, void*,IScriptEnvironment* env) {
 AVSValue __cdecl Create_FadeOut(AVSValue args, void*,IScriptEnvironment* env) {
   const int duration = args[1].AsInt();
   const int fadeclr = args[2].AsInt(0);
-  const float fps = args[3].AsFloat(24);
+  const float fps = (float)args[3].AsFloat(24);
   PClip a = args[0].AsClip();
   PClip b = ColorClip(a,duration+1,fadeclr,fps,env);
   return new Dissolve(a, b, duration, fps, env);
@@ -886,7 +886,7 @@ AVSValue __cdecl Create_FadeOut(AVSValue args, void*,IScriptEnvironment* env) {
 AVSValue __cdecl Create_FadeOut2(AVSValue args, void*,IScriptEnvironment* env) {
   const int duration = args[1].AsInt();
   const int fadeclr = args[2].AsInt(0);
-  const float fps = args[3].AsFloat(24);
+  const float fps = (float)args[3].AsFloat(24);
   PClip a = args[0].AsClip();
   PClip b = ColorClip(a,duration+2,fadeclr,fps,env);
   return new Dissolve(a, b, duration, fps, env);
@@ -895,7 +895,7 @@ AVSValue __cdecl Create_FadeOut2(AVSValue args, void*,IScriptEnvironment* env) {
 AVSValue __cdecl Create_FadeIn0(AVSValue args, void*,IScriptEnvironment* env) {
   const int duration = args[1].AsInt();
   const int fadeclr = args[2].AsInt(0);
-  const float fps = args[3].AsFloat(24);
+  const float fps = (float)args[3].AsFloat(24);
   PClip a = args[0].AsClip();
   PClip b = ColorClip(a,duration,fadeclr,fps,env);
   return new Dissolve(b, a, duration, fps, env);
@@ -904,7 +904,7 @@ AVSValue __cdecl Create_FadeIn0(AVSValue args, void*,IScriptEnvironment* env) {
 AVSValue __cdecl Create_FadeIn(AVSValue args, void*,IScriptEnvironment* env) {
   const int duration = args[1].AsInt();
   const int fadeclr = args[2].AsInt(0);
-  const float fps = args[3].AsFloat(24);
+  const float fps = (float)args[3].AsFloat(24);
   PClip a = args[0].AsClip();
   PClip b = ColorClip(a,duration+1,fadeclr,fps,env);
   return new Dissolve(b, a, duration, fps, env);
@@ -913,7 +913,7 @@ AVSValue __cdecl Create_FadeIn(AVSValue args, void*,IScriptEnvironment* env) {
 AVSValue __cdecl Create_FadeIn2(AVSValue args, void*,IScriptEnvironment* env) {
   const int duration = args[1].AsInt();
   const int fadeclr = args[2].AsInt(0);
-  const float fps = args[3].AsFloat(24);
+  const float fps = (float)args[3].AsFloat(24);
   PClip a = args[0].AsClip();
   PClip b = ColorClip(a,duration+2,fadeclr,fps,env);
   return new Dissolve(b, a, duration, fps, env);
@@ -923,7 +923,7 @@ AVSValue __cdecl Create_FadeIO0(AVSValue args, void*, IScriptEnvironment* env) {
 	try {	// HIDE DAMN SEH COMPILER BUG!!!
   const int duration = args[1].AsInt();
   const int fadeclr = args[2].AsInt(0);
-  const float fps = args[3].AsFloat(24);
+  const float fps = (float)args[3].AsFloat(24);
   PClip a = args[0].AsClip();
   PClip b = ColorClip(a,duration,fadeclr,fps,env);
   AVSValue dissolve_args[] = { b, a, b, duration, fps };
@@ -936,7 +936,7 @@ AVSValue __cdecl Create_FadeIO(AVSValue args, void*, IScriptEnvironment* env) {
 	try {	// HIDE DAMN SEH COMPILER BUG!!!
   const int duration = args[1].AsInt();
   const int fadeclr = args[2].AsInt(0);
-  const float fps = args[3].AsFloat(24);
+  const float fps = (float)args[3].AsFloat(24);
   PClip a = args[0].AsClip();
   PClip b = ColorClip(a,duration+1,fadeclr,fps,env);
   AVSValue dissolve_args[] = { b, a, b, duration, fps };
@@ -949,7 +949,7 @@ AVSValue __cdecl Create_FadeIO2(AVSValue args, void*, IScriptEnvironment* env) {
 	try {	// HIDE DAMN SEH COMPILER BUG!!!
   const int duration = args[1].AsInt();
   const int fadeclr = args[2].AsInt(0);
-  const float fps = args[3].AsFloat(24);
+  const float fps = (float)args[3].AsFloat(24);
   PClip a = args[0].AsClip();
   PClip b = ColorClip(a,duration+2,fadeclr,fps,env);
   AVSValue dissolve_args[] = { b, a, b, duration, fps };
